[clear] faIconを別ファイルに移動
